progress on merkle refactor
